+++ src/pink/delays.clj
@@ -1,6 +1,6 @@
 (ns pink.delays
   (:require [pink.config :refer :all]
-            [pink.util :refer [create-buffer mix-buffers]]
+            [pink.util :refer [create-buffer mix-buffers generator]]
             ))
 
 ;; feedback functions
@@ -54,25 +54,63 @@
       (System/arraycopy delay-buffer read-ptr out 0 buf-size)
       new-end))))
 
+(defn samp-delay
+  "Non-interpolating delay-line with fixed-delay-time. delay-time given in
+  samples."
+  [afn ^long delay-time]
+  (let [out (create-buffer) 
+        ^doubles delay-buffer (double-array delay-time)
+        rw-ptr (int-array 1 0)]
+    (generator
+      [rw-ptr (long 0)]
+      [sig afn]
+      (let [v (aget delay-buffer rw-ptr)]
+        (aset delay-buffer rw-ptr sig)
+        (aset out int-indx v)
+        (recur (unchecked-inc indx) 
+               (rem (unchecked-inc rw-ptr) delay-time)))
+      (yield out))))
+
+;(defn frac-delay
+;  "Fixed length, interpolating delay-line. delay-time given in samples."
+;  [afn ^double delay-time]
+;  (generator
+;      [rw-ptr (long 0)]
+;      [sig afn]
+;      (let [v (aget delay-buffer rw-ptr)]
+;        (aset delay-buffer rw-ptr sig)
+;        (aset out int-indx v)
+;        (recur (unchecked-inc indx) 
+;               (rem (unchecked-inc rw-ptr) delay-time)))
+;      (yield out)))
+
+
 (defn adelay
-  "Fixed length, non-interpolating delay-line. delay-time given in seconds."
+  "Non-interpolating delay-line with fixed delay-time. delay-time given in
+  seconds."
   [afn ^double delay-time]
-  (let [out ^doubles (create-buffer) 
-        delay-buffer-len (int (+ 0.5 (* delay-time (double *sr*))))
-        delay-buffer (double-array delay-buffer-len 0.0)
-        read-ptr (int-array 1 *buffer-size*) ; start reading one buffer ahead
-        write-ptr (int-array 1 0)]
-    (fn []
-      (when-let [asig ^doubles (afn)]
-        (let [cur-read-ptr ^int (aget read-ptr 0)
-              cur-write-ptr ^int (aget write-ptr 0)
-              new-write-ptr
-              (do-write asig delay-buffer cur-write-ptr delay-buffer-len)
-              new-read-ptr 
-              (do-read out delay-buffer cur-read-ptr delay-buffer-len)]
-          (aset read-ptr 0 (int new-read-ptr))
-          (aset write-ptr 0 (int new-write-ptr))
-          out)))))
+  (samp-delay afn (int (+ 0.5 (* delay-time (double *sr*))))))
+
+;(defn fdelay
+;  "Fractional delay. Time is given in seconds."
+;  [afn ^double delay-time]
+;  (let [out ^doubles (create-buffer) 
+;        delay-buffer-len (int (+ 0.5 (* delay-time (double *sr*))))
+;        delay-buffer (double-array delay-buffer-len 0.0)
+;        read-ptr (int-array 1 *buffer-size*) ; start reading one buffer ahead
+;        write-ptr (int-array 1 0)]
+;    (fn []
+;      (when-let [asig ^doubles (afn)]
+;        (let [cur-read-ptr ^int (aget read-ptr 0)
+;              cur-write-ptr ^int (aget write-ptr 0)
+;              new-write-ptr
+;              (do-write asig delay-buffer cur-write-ptr delay-buffer-len)
+;              new-read-ptr 
+;              (do-read out delay-buffer cur-read-ptr delay-buffer-len)]
+;          (aset read-ptr 0 (int new-read-ptr))
+;          (aset write-ptr 0 (int new-write-ptr))
+;          out))))
+;  )
 
 
 ;; Multi-Tap Delay